Inject AnyScheduler using `PowerfulCombine`
